добавит модуль создания массивов фото
